Syntax update to support different versions of Node
Some of the community distributions of Linux on the
BeagleBone Black include an older version of Node than the
Linux Processor SDK. It was found that the original syntax
for the AWS Thing Shadow update used in this package did
not work with the older version of Node.

This commit changes the syntax so that the demo will run on
both the TI Linux Processor SDK version of Node as well as
the BeagleBoard community version of Node.

Signed-off-by: Jason Reeder <jreeder@ti.com>